Moved the creation of the Permission List Box to the Sharing Permissions Grid. Un-gxt'ed the ACL Editor.
