+++ src/main/java/org/sagebionetworks/web/client/widget/sharing/EvaluationAccessControlListEditorViewImpl.java
@@ -1,6 +1,8 @@
 package org.sagebionetworks.web.client.widget.sharing;
 
+import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 import org.sagebionetworks.web.client.DisplayConstants;
@@ -16,43 +18,43 @@ import org.sagebionetworks.web.shared.PublicPrincipalIds;
 import org.sagebionetworks.web.shared.users.AclEntry;
 import org.sagebionetworks.web.shared.users.PermissionLevel;
 
-import com.extjs.gxt.ui.client.data.ModelData;
+import com.extjs.gxt.ui.client.store.ListStore;
+import com.extjs.gxt.ui.client.widget.grid.ColumnConfig;
+import com.extjs.gxt.ui.client.widget.grid.ColumnData;
+import com.extjs.gxt.ui.client.widget.grid.ColumnModel;
+import com.extjs.gxt.ui.client.widget.grid.Grid;
+import com.extjs.gxt.ui.client.widget.grid.GridCellRenderer;
 import com.extjs.gxt.ui.client.event.ButtonEvent;
 import com.extjs.gxt.ui.client.event.Events;
 import com.extjs.gxt.ui.client.event.GridEvent;
 import com.extjs.gxt.ui.client.event.Listener;
 import com.extjs.gxt.ui.client.event.MenuEvent;
-import com.extjs.gxt.ui.client.event.MessageBoxEvent;
 import com.extjs.gxt.ui.client.event.SelectionChangedEvent;
 import com.extjs.gxt.ui.client.event.SelectionChangedListener;
 import com.extjs.gxt.ui.client.event.SelectionListener;
-import com.extjs.gxt.ui.client.store.ListStore;
 import com.extjs.gxt.ui.client.widget.BoxComponent;
-import com.extjs.gxt.ui.client.widget.Dialog;
 import com.extjs.gxt.ui.client.widget.HorizontalPanel;
 import com.extjs.gxt.ui.client.widget.Label;
 import com.extjs.gxt.ui.client.widget.LayoutContainer;
-import com.extjs.gxt.ui.client.widget.MessageBox;
 import com.extjs.gxt.ui.client.widget.button.Button;
-import com.extjs.gxt.ui.client.widget.form.ComboBox;
 import com.extjs.gxt.ui.client.widget.form.ComboBox.TriggerAction;
 import com.extjs.gxt.ui.client.widget.form.FieldSet;
 import com.extjs.gxt.ui.client.widget.form.FormPanel;
 import com.extjs.gxt.ui.client.widget.form.SimpleComboBox;
 import com.extjs.gxt.ui.client.widget.form.SimpleComboValue;
-import com.extjs.gxt.ui.client.widget.grid.ColumnData;
-import com.extjs.gxt.ui.client.widget.grid.ColumnModel;
-import com.extjs.gxt.ui.client.widget.grid.Grid;
-import com.extjs.gxt.ui.client.widget.grid.GridCellRenderer;
 import com.extjs.gxt.ui.client.widget.layout.FlowLayout;
 import com.extjs.gxt.ui.client.widget.layout.FormLayout;
 import com.extjs.gxt.ui.client.widget.layout.MarginData;
 import com.extjs.gxt.ui.client.widget.layout.TableData;
 import com.extjs.gxt.ui.client.widget.menu.Menu;
 import com.extjs.gxt.ui.client.widget.menu.MenuItem;
+import com.google.gwt.event.dom.client.ClickEvent;
+import com.google.gwt.event.dom.client.ClickHandler;
+import com.google.gwt.resources.client.ImageResource;
 import com.google.gwt.safehtml.shared.SafeHtmlUtils;
 import com.google.gwt.user.client.Element;
 import com.google.gwt.user.client.ui.AbstractImagePrototype;
+import com.google.gwt.user.client.ui.Anchor;
 import com.google.gwt.user.client.ui.HTML;
 import com.google.gwt.user.client.ui.Widget;
 import com.google.inject.Inject;
@@ -65,6 +67,8 @@ public class EvaluationAccessControlListEditorViewImpl extends LayoutContainer i
 	static final String REMOVE_COLUMN_ID = "removeData";
 	private static final int DEFAULT_WIDTH = 380;
 	private static final int BUTTON_PADDING = 3;
+	
+	private static final String STYLE_VERTICAL_ALIGN_MIDDLE = "vertical-align:middle !important;";
 	
 	private Presenter presenter;
 	private IconsImageBundle iconsImageBundle;
@@ -158,11 +162,11 @@ public class EvaluationAccessControlListEditorViewImpl extends LayoutContainer i
 
 		// show existing permissions
 		permissionsStore = new ListStore<PermissionsTableEntry>();
-		permissionsGrid = AccessControlListEditorViewImpl.createPermissionsGrid(
+		permissionsGrid = createPermissionsGrid(
 				permissionsStore, 
-				AccessControlListEditorViewImpl.createPeopleRenderer(publicPrincipalIds, synapseJSNIUtils, iconsImageBundle), 
+				createPeopleRenderer(publicPrincipalIds, synapseJSNIUtils, iconsImageBundle), 
 				createButtonRenderer(),
-				AccessControlListEditorViewImpl.createRemoveRenderer(iconsImageBundle, new CallbackP<Long>() {
+				createRemoveRenderer(iconsImageBundle, new CallbackP<Long>() {
 					@Override
 					public void invoke(Long principalId) {
 						presenter.removeAccess(principalId);
@@ -432,6 +436,126 @@ public class EvaluationAccessControlListEditorViewImpl extends LayoutContainer i
 		} else {
 			showAddMessage("Please select a user or group to grant permission to.");
 		}
+	}
+	
+	public Grid<PermissionsTableEntry> createPermissionsGrid(
+			ListStore<PermissionsTableEntry> permissionsStore,
+			GridCellRenderer<PermissionsTableEntry> peopleRenderer,
+			GridCellRenderer<PermissionsTableEntry> buttonRenderer,
+			GridCellRenderer<PermissionsTableEntry> removeRenderer,
+			boolean isEditable) {
+		List<ColumnConfig> configs = new ArrayList<ColumnConfig>();
+
+		ColumnConfig column = new ColumnConfig();
+		column.setId(PRINCIPAL_COLUMN_ID);
+		column.setHeader("People");
+		column.setWidth(200);
+		column.setRenderer(peopleRenderer);
+		configs.add(column);
+
+		column = new ColumnConfig();
+		column.setId(ACCESS_COLUMN_ID);
+		column.setHeader("Access");
+		column.setWidth(110);
+		column.setRenderer(buttonRenderer);
+		column.setStyle(STYLE_VERTICAL_ALIGN_MIDDLE);
+		configs.add(column);
+
+		column = new ColumnConfig();
+		column.setId(REMOVE_COLUMN_ID);
+		column.setHeader("");
+		column.setWidth(25);
+		column.setRenderer(removeRenderer);
+		column.setStyle(STYLE_VERTICAL_ALIGN_MIDDLE);
+		column.setHidden(!isEditable);
+		configs.add(column);
+
+		Grid<PermissionsTableEntry> permissionsGrid = new Grid<PermissionsTableEntry>(
+				permissionsStore, new ColumnModel(configs));
+		permissionsGrid.setAutoExpandColumn(PRINCIPAL_COLUMN_ID);
+		permissionsGrid.setBorders(true);
+		permissionsGrid.setWidth(520);
+		permissionsGrid.setHeight(180);
+		return permissionsGrid;
+	}
+	
+	public GridCellRenderer<PermissionsTableEntry> createPeopleRenderer(
+			final PublicPrincipalIds publicPrincipalIds, 
+			final SynapseJSNIUtils synapseJSNIUtils,
+			final IconsImageBundle iconsImageBundle) {
+		GridCellRenderer<PermissionsTableEntry> personRenderer = new GridCellRenderer<PermissionsTableEntry>() {
+			@Override
+			public Object render(PermissionsTableEntry model, String property,
+					ColumnData config, int rowIndex, int colIndex,
+					ListStore<PermissionsTableEntry> store,
+					Grid<PermissionsTableEntry> grid) {
+				PermissionsTableEntry entry = store.getAt(rowIndex);
+				AclEntry aclEntry = entry.getAclEntry();
+				String principalHtml = "";
+				Long publicPrincipalId = publicPrincipalIds.getPublicAclPrincipalId();
+				Long authenticatedPrincipalId = publicPrincipalIds.getAuthenticatedAclPrincipalId();
+				Long anonymousUserPrincipalId = publicPrincipalIds.getAnonymousUserPrincipalId();
+				
+				if (aclEntry != null & aclEntry.getOwnerId() != null) {
+					if (publicPrincipalId != null && aclEntry.getOwnerId().equals(publicPrincipalId.toString())) {
+						//is public group
+						principalHtml = DisplayUtils.getUserNameDescriptionHtml(DisplayConstants.PUBLIC_ACL_TITLE, DisplayConstants.PUBLIC_ACL_DESCRIPTION);
+					} else if (authenticatedPrincipalId != null && aclEntry.getOwnerId().equals(authenticatedPrincipalId.toString())) {
+						//is authenticated group
+						principalHtml = DisplayUtils.getUserNameDescriptionHtml(DisplayConstants.AUTHENTICATED_USERS_ACL_TITLE, DisplayConstants.AUTHENTICATED_USERS_ACL_DESCRIPTION);	
+					} else if (anonymousUserPrincipalId != null && aclEntry.getOwnerId().equals(anonymousUserPrincipalId.toString())) {
+						//is anonymous user
+						principalHtml = DisplayUtils.getUserNameDescriptionHtml(DisplayConstants.PUBLIC_USER_ACL_TITLE, DisplayConstants.PUBLIC_USER_ACL_DESCRIPTION);
+					} else {
+						principalHtml = DisplayUtils.getUserNameDescriptionHtml(aclEntry.getTitle(), aclEntry.getSubtitle());
+					}
+				}
+				
+				String iconHtml = "";
+				if (publicPrincipalId != null && aclEntry.getOwnerId().equals(publicPrincipalId.toString())){
+					ImageResource icon = iconsImageBundle.globe32();
+					iconHtml = DisplayUtils.getIconThumbnailHtml(icon);	
+				} else if (!aclEntry.isIndividual()) {
+					//if a group, then try to fill in the icon from the team
+					String url = DisplayUtils.createTeamIconUrl(
+							synapseJSNIUtils.getBaseFileHandleUrl(), 
+							aclEntry.getOwnerId()
+					);
+					iconHtml = DisplayUtils.getThumbnailPicHtml(url);
+				} else {
+					// try to get the userprofile picture
+					String url = DisplayUtils.createUserProfilePicUrl(
+							synapseJSNIUtils.getBaseProfileAttachmentUrl(), 
+							aclEntry.getOwnerId() 
+					);
+					iconHtml = DisplayUtils.getThumbnailPicHtml(url);
+				}
+				return iconHtml + "&nbsp;&nbsp;" + principalHtml;
+			}
+			
+		};
+		return personRenderer;
+	}
+
+	public GridCellRenderer<PermissionsTableEntry> createRemoveRenderer(final IconsImageBundle iconsImageBundle, final CallbackP<Long> callback) {
+		GridCellRenderer<PermissionsTableEntry> removeButton = new GridCellRenderer<PermissionsTableEntry>() {  			   
+			@Override  
+			public Object render(final PermissionsTableEntry model, String property, ColumnData config, int rowIndex,  
+				  final int colIndex, ListStore<PermissionsTableEntry> store, Grid<PermissionsTableEntry> grid) {				 
+				  final PermissionsTableEntry entry = store.getAt(rowIndex);
+					Anchor removeAnchor = new Anchor();
+					removeAnchor.setHTML(DisplayUtils.getIconHtml(iconsImageBundle.deleteButton16()));
+					removeAnchor.addClickHandler(new ClickHandler() {			
+						@Override
+						public void onClick(ClickEvent event) {
+							Long principalId = (Long.parseLong(entry.getAclEntry().getOwnerId()));
+							callback.invoke(principalId);
+						}
+					});
+					return removeAnchor;
+			  }
+			};  
+		return removeButton;
 	}
 }
